Fix Story integration test, iteration date serialization, and split integration tests per resource
Story.workflow/Story.state changed shape (workflow now returns the parent
Workflow, state is a getter resolving to a WorkflowState), so
story-lifecycle.integration.test.ts no longer matched the API and failed to
compile. Updated the assertions to match.

Iteration creation 422'd because BaseResource serialized Date fields as full
ISO timestamps, which Shortcut's /iterations endpoint rejects for
start_date/end_date. Added an opt-in dateOnlyFields hook on BaseResource and
declared it on Iteration. That in turn exposed a second bug: convertApiFields
never parsed the API's bare YYYY-MM-DD dates back into Date objects, since
isValidDatetimeFormat required a full datetime; loosened the regex to accept
a date-only string.

Also split integration/resource.integration.test.ts into one file per
resource (label, epic, iteration, workflow, member), matching the existing
per-resource convention already used by story.integration.test.ts.
