+++ tests/iterations/iteration.test.ts
@@ -1,7 +1,11 @@
+import { AxiosInstance } from 'axios'
+
 import Iteration from '@sx/iterations/iteration'
 import Label from '@sx/labels/label'
 import Team from '@sx/teams/team'
 import TeamsService from '@sx/teams/teams-service'
+
+import { stubHttp } from '../helpers/http'
 
 
 describe('Iteration class', () => {
@@ -50,5 +54,20 @@ describe('Iteration class', () => {
     const iteration = new Iteration({})
     expect(iteration.createFields).toEqual(expect.arrayContaining(expectedFields))
     expect(iteration.createFields.length).toBe(expectedFields.length)
+  })
+
+  it('sends start_date and end_date as date-only strings on create', async () => {
+    const http: AxiosInstance = stubHttp()
+    const post = http.post as jest.Mock
+    post.mockResolvedValue({data: {id: 1}})
+    const startDate = new Date('2026-07-29T12:31:07.768Z')
+    const endDate = new Date('2026-08-05T12:31:07.768Z')
+    const iteration = new Iteration({name: 'Iteration 1', startDate, endDate}).setHttp(http)
+
+    await iteration.save()
+
+    const [, body] = post.mock.calls[0] as [string, { start_date: string; end_date: string }]
+    expect(body.start_date).toBe('2026-07-29')
+    expect(body.end_date).toBe('2026-08-05')
   })
 })
